Practiced Persistent Volumes & Perssitent Volume Claims along with static & dynamic provisioning
For static provisioning I created a persistent volume using my own local storage. Then created a persistent volume claim to be attached to it.
Then created a deployment in which a container mounts the storage.
For dynamic provisioning I practiced with storage classes and how to provision storage with a sample deployment.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>